requirements.txt
requests 
django-redis